backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
  "fastapi==0.115.12",
  "uvicorn==0.34.0",
  "pydantic[email] (>=2.11.1,<3.0.0)",
  "typing-inspect==0.9.0",
  "typing-extensions==4.13.0",
  "pymongo==4.11.3",
  "motor==3.7.0",
  "beanie==1.29.0",
  "openai==1.69.0",
  "python-jose==3.4.0",
  "passlib==1.7.4",
  "python-multipart==0.0.6",
  "bcrypt==4.0.1",
  "pyjwt==2.6.0",
  "google-auth==2.38.0",
  "google-auth-oauthlib==1.2.1",
  "google-auth-httplib2==0.2.0",
  "google-api-python-client==2.166.0",
  "langchain==0.3.21",
  "langchain-openai==0.3.11",
  "langchain-core==0.3.49",
  "langchain-community==0.3.20",
  "langchain-text-splitters==0.3.7",
  "azure-ai-inference==1.0.0b9",
  "pydantic-core==2.33.0",
  "python-dotenv==1.1.0",
  "pytest==7.2.2",
  "requests==2.28.2",
  "streamlit==1.44.0",
  "PyMuPDF==1.25.4",
  "jinja2==3.1.2",
  "plotly (>=6.0.1,<7.0.0)",
  "spacy==3.8.4",
  "scikit-learn==1.6.1",
  "cryptography (>=44.0.2,<45.0.0)",
  "bs4 (>=0.0.2,<0.0.3)",
  "pdfkit (>=1.0.0,<2.0.0)",
  "docx (>=0.2.4,<0.3.0)",
  "markdown (>=3.7,<4.0)",
  "msgraph-core (>=1.3.3,<2.0.0)",
  "azure-identity (>=1.21.0,<2.0.0)",
  "beautifulsoup4 (>=4.13.3,<5.0.0)",
  "msal (>=1.32.0,<2.0.0)",
  "numpy>=2.2.6",
]

[tool.uv.sources]
a2a = { path = "./src/", editable = true }
